Move the remote-tab rebuild thunk into its own workflow module
sessionThunks.ts crossed the 500-line lint cap and the generated binding's
non-null typing made the runtime null-default look unnecessary to the
linter. The tab-rebuild flow gets its own module (composing the same slice
actions trackOpenSessionMetadata dispatches), and the binding result is
widened to string[] | null before defaulting — Go nil arrives as null.
