+++ erun-ui/frontend/src/app/sessionThunks.ts
@@ -2,7 +2,6 @@ import type { StartSessionResult, UISelection } from '@/types';
 
 import {
   CloseSession,
-  ListRemoteAppSessions,
   StartAISession,
   StartDeploySession,
   StartInitSession,
@@ -13,6 +12,7 @@ import { resolveAutoStartGate } from './autoStartGate';
 import { applyPendingDebugHeader, setPendingDebugHeader, syncDebugDisplay } from './debugThunks';
 import { readError } from './errors';
 import { hideTerminalMessage, showTerminalMessage } from './notificationThunks';
+import { reattachRemoteTerminalTabs } from './remoteSessionTabsThunks';
 import { loadReviewDiff } from './reviewThunks';
 import { selectActiveSlotForSelection, selectEnvironmentExists } from './selectors';
 import { isNewSessionSelection } from './sessionSelection';
@@ -166,41 +166,6 @@ export const ensureDefaultEnvTabs =
     await dispatch(ensureLiveDefaultTab(key, runSelection, 'erun', 'ERun', cols, rows));
     await dispatch(ensureLiveDefaultTab(key, runSelection, 'local', 'Local', cols, rows));
     await dispatch(ensureLiveDefaultTab(key, runSelection, 'ai', 'AI', cols, rows));
-  };
-
-// reattachRemoteTerminalTabs rebuilds tabs for persistent pod sessions this
-// window does not know about — custom terminals (`open-N`) created in another
-// ERun window or a previous run that are still running in the pod. The default
-// tabs attach-or-create on their own in ensureDefaultEnvTabs; this fills in
-// the extras, so every running remote session ends up attached here (the
-// attach itself takes the session over from any other window). Detection is
-// best-effort and must never stall the open flow.
-const reattachRemoteTerminalTabs =
-  (runSelection: UISelection, key: string, cols: number, rows: number): AppThunk<Promise<void>> =>
-  async (dispatch, getState) => {
-    let ids: string[];
-    try {
-      ids = ((await ListRemoteAppSessions(runSelection)) ?? []) as string[];
-    } catch {
-      return; // env may be local-only, unreachable, or not deployed
-    }
-    for (const id of ids) {
-      const match = /^open-([1-9]\d*)$/.exec(id);
-      if (!match) {
-        continue;
-      }
-      const slot = Number(match[1]);
-      const tabs = getState().terminal.tabsByEnv[key] ?? [];
-      if (tabs.some((tab) => tab.slot === slot && tab.kind === 'extra')) {
-        continue;
-      }
-      try {
-        const result = (await StartSession(runSelection, slot, cols, rows)) as StartSessionResult;
-        dispatch(trackOpenSessionMetadata(key, result, runSelection));
-      } catch {
-        // skip sessions that fail to attach; the next env open retries
-      }
-    }
   };
 
 // surfaceEnvSession repoints the visible terminal at the new env's
